1st commit
This case study dives into the challenges and solutions
associated with building a high-performance E-commerce Sales Transaction system for Jiggy
E-commerce. By employing Python for data manipulation, SQL for data querying, and PostgreSQL as
the underlying database, the aim is to construct a system that stands at the forefront of efficiency, scalability, and reliability. h. The ultimate vision is to transform Jiggy E-commerce into a data-savvy
enterprise, where every transaction is an opportunity for optimization and customer delight.